backend/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
labelimg==1.8.6
lxml==5.4.0
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pyqt5==5.15.11
pyqt5-qt5==5.15.17
pyqt5-sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.145
ultralytics-thop==2.0.14
urllib3==2.4.0
